Switch to Layout Preference / Adjust padding
This fixes a very weird issue that was not noticable until the DPI
was increased and we started to swip back and forth. What we noticed
was that at the end of the preference a very small grey box would appear
for 2-3 seconds. The issue is not present if there is not a custom layout
attached to the preference.

Switching to LayoutPreference fixes this issue and brings us inline with
how Google has handled preferences with custom layouts.

Adjusted the bottom margin on layouts to make up the difference in the bottom
text touching the preference under it. Anyways, #BlameNathan lol he found this
issue...

- Adjust padding some more, last time hopefully :p

Change-Id: Ife6e909ce2369e935173c2abe7f35b0a6bf39d97